requirements.txt
async-lru==1.0.2
dnspython==2.1.0
lxml==4.6.2
pyaes==1.6.1
Pyrogram==1.1.13
PySocks==1.7.1
python-docx==0.8.10
TgCrypto==1.2.2
